# Add optimized reduce(vcat, dataframes)