--- a/test/cat.jl
+++ b/test/cat.jl
@@ -293,6 +293,6 @@ end
 end
 x = view(DataFrame(A = Vector{Union{Missing, Int}}(1:3)), 2:2, :)
 y = DataFrame(A = 4:5)
-@test vcat(x, y) == DataFrame(A = [2, 4, 5]) == reduce(vcat, [x,y])
+@test vcat(x, y) == DataFrame(A = [2, 4, 5]) == reduce(vcat, [x, y])
 
 end # module