https://github.com/JuliaLang/julia/issues/21672
makes this the preferred phrasing for `vcat`ing collections of matrixes.
it is nice for it to be the same for dataframes.

this very thinly wraps the existing `_vcat`, see the timing.

```
julia> using BenchmarkTools

julia> dataframes = [DataFrame(;A=1:100, B=rand(100), C=fill(i,100)) for i in 1:50];

julia> @btime vcat($(dataframes)...);
  37.223 μs (422 allocations: 138.52 KiB)

julia> @btime reduce(vcat, $(dataframes));
  595.878 μs (4904 allocations: 3.21 MiB)

julia> @btime DataFrames._vcat($(dataframes));
  36.184 μs (420 allocations: 137.59 KiB)

julia> dataframes = [DataFrame(;A=1:100, B=rand(100), C=fill(i,100)) for i in 1:500];

julia> @btime vcat($(dataframes)...);
  306.130 μs (3572 allocations: 1.31 MiB)

julia> @btime DataFrames._vcat($(dataframes));
  298.962 μs (3570 allocations: 1.30 MiB)

julia> dataframes = [DataFrame(;A=1:100, B=rand(100), C=fill(i,100)) for i in 1:5000];

julia> @btime vcat($(dataframes)...);
  6.841 ms (62016 allocations: 13.50 MiB)

julia> @btime DataFrames._vcat($(dataframes));
  6.411 ms (62011 allocations: 13.35 MiB)

julia> dataframes = [DataFrame(;A=1:100, B=rand(100), C=fill(i,100)) for i in 1:500_000];

julia> @btime vcat($(dataframes)...);
  3.216 s (6497016 allocations: 1.32 GiB)

julia> @btime DataFrames._vcat($(dataframes));
  3.528 s (6497011 allocations: 1.31 GiB)
```

Only a little faster than splatting.
The splatting penalty is pretty low for this vs that actual time involved.