@bkamins 2019-02-21
Yes, from my tests the benefit is no more than ~10%. But the implementation is simple, so I think it does not hurt to have it 😄. Thank you.